[magic] Fix 32-bit magic instruction implementation. Use register GDX as the second register, not GBX to be compatible with -fPIC. Fix 32-bit Sim(Set|Get)OwnFreqMHz

+++ test/dvfs/dvfs.c
@@ -23,23 +23,23 @@ int main() {
 	value = 0;
 	loop();
 
-	unsigned long long freq = SimGetOwnFreqMHz();
-	printf("Current Freq = %lld MHz\n", freq);
+	unsigned long freq = SimGetOwnFreqMHz();
+	printf("Current Freq = %lu MHz\n", freq);
 
 	freq = 5000;
-	printf("Setting frequency to %lld MHz\n", freq);
+	printf("Setting frequency to %lu MHz\n", freq);
 	SimSetOwnFreqMHz(freq);
 	freq = SimGetOwnFreqMHz();
-	printf("Current Freq = %lld MHz\n", freq);
+	printf("Current Freq = %lu MHz\n", freq);
 
 	value = 0;
 	loop();
 
 	freq = 1000;
-	printf("Setting frequency to %lld MHz\n", freq);
+	printf("Setting frequency to %lu MHz\n", freq);
 	SimSetFreqMHz(1, freq);
 	freq = SimGetFreqMHz(1);
-	printf("Current Freq = %lld MHz\n", freq);
+	printf("Current Freq = %lu MHz\n", freq);
 
 	value = 0;
 	loop();
